Refactor open redirect scanner with enhanced features and reporting
