+++ ultimate_analysis.csv
@@ -1,0 +1,1 @@
+name,value,source,context,url,is_redirect_related,confidence,vulnerability_found
