+++ ultimate_results.json
@@ -1,0 +1,13 @@
+{
+  "scan_info": {
+    "target_url": "https://httpbin.org/redirect-to",
+    "scan_date": "2025-09-03T22:04:09.161797",
+    "scanner_version": "Ultimate Redirect Hunter v3.0",
+    "total_parameters": 0,
+    "redirect_parameters": 0,
+    "vulnerabilities_found": 0,
+    "web3_detected": false
+  },
+  "parameters": [],
+  "vulnerabilities": []
+}
